move json marshalling to a dedicated file
