+++ qbft/instance.go
@@ -220,41 +220,3 @@ func (i *Instance) Decode(data []byte) error {
 func (i *Instance) CanProcessMessages() bool {
 	return !i.forceStop && int(i.State.Round) < CutoffRound
 }
-
-/////////////////////// JSON Marshalling for Tests ///////////////////////
-
-func (i *Instance) MarshalJSON() ([]byte, error) {
-	type Alias Instance
-	if i.forceStop {
-		return json.Marshal(&struct {
-			ForceStop bool `json:"forceStop"`
-			*Alias
-		}{
-			ForceStop: i.forceStop,
-			Alias:     (*Alias)(i),
-		})
-	} else {
-		return json.Marshal(&struct {
-			*Alias
-		}{
-			Alias: (*Alias)(i),
-		})
-	}
-}
-
-func (i *Instance) UnmarshalJSON(data []byte) error {
-	type Alias Instance
-	aux := &struct {
-		ForceStop *bool `json:"forceStop,omitempty"`
-		*Alias
-	}{
-		Alias: (*Alias)(i),
-	}
-	if err := json.Unmarshal(data, &aux); err != nil {
-		return err
-	}
-	if aux.ForceStop != nil {
-		i.forceStop = *aux.ForceStop
-	}
-	return nil
-}
